Moved code into clean Rails 7 install
No git history thru Ch 40:
- Attempt to upgrade the existing Rails 6 coursework to Rails 7
(instead of re-doing completed lessons in a clean install) was
problematic because Rails 7 doesn't use webpack. Wasn't even aware
prev Rails used webpack and didn't know this was an issue
until having issues attempting to deploy it. Seems happy now, but
bummer to loose the git history here. That history is in tact in the
other build tho.
- Another thing to note: using the `seed_dump` gem to backup the
local db, but had to make a small fix to it to make it work.
It's an old repo and my PR won't likely get merged, so i'm linking
to my fork of it.